Day 7

+++ IPuzzle.cs
@@ -3,8 +3,8 @@
 public interface IPuzzle
 {
     void LoadInput(string inputPath);
-    int Solve(string inputPath, int part) { LoadInput(inputPath); return SolvePart(part); }
-    int SolvePart(int part) => part == 1 ? Part1() : Part2();
-    int Part1() => int.MinValue;
-    int Part2() => int.MinValue;
+    long Solve(string inputPath, int part) { LoadInput(inputPath); return SolvePart(part); }
+    long SolvePart(int part) => part == 1 ? Part1() : Part2();
+    long Part1() => long.MinValue;
+    long Part2() => long.MinValue;
 }
